+++ HISTORY.rst
@@ -6,7 +6,9 @@ History
 0.3.0 (2014-05-??)
 ++++++++++++++++++
 
+* Added extensible Senderable class to expedite creating new senders.
 * Added async_redis sender
+* Added travis-ci
 
 
 0.2.0 (2014-05-13)